Create a new graphData selector to allow storing server response without changes
